refactor(test): deduplicate export round-trip blocks across pt_expt tests
- Add nf=1 slice verification to export_save_load_and_compare in export_helpers.py
- Add model_forward_lower_export_round_trip helper for model-level export tests
- Remove ~14-line "Different shapes" blocks from 9 descriptor and 6 fitting tests
- Replace ~60-100 line test bodies in 6 model tests with helper calls
- Fix race condition in test_export_pipeline.py temp file handling
- Merge test_linear_ener_model.py into test_linear_model.py (same module)
